Implement recursive scan.
This means we can deal with arrays of arbitrary size where the (intermediate) partials array itself requires a scan operation.
Previously, we assumed/limited the size of the partials array so that we could always do the scan with one workgroup.
Now we can just use the stack to keep hold of all the intermediate data :-)